# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 食堂饭卡管理系统 API
  description: |
    食堂饭卡管理系统后端接口契约，前后端严格遵照此文档开发。

    **金额单位**：统一为分（int64）。
    **卡号（cardNo）**：16 位数字字符串，独立于数据库自增 ID，所有接口路径参数均使用 cardNo。
    **押金**：固定 2000 分（20 元），系统常量，不由前端传入。
  version: 2.0.0
servers:
  - url: http://localhost:8080
    description: 主后端（开发环境）
  - url: http://localhost:9090
    description: 学籍验证 Mock 服务（开发环境）
...
paths:
...
  /api/cards/{cardNo}/deposits:
    post:
      tags: [deposits]
      summary: 存款（充值）
      description: |
        为指定卡号充值。充值成功后返回收据信息（卡号、持卡人姓名、充值金额、充值后余额、充值时间）。

        约束：
        - 充值金额必须 > 0
        - 卡必须为 active 状态
      operationId: deposit
      parameters:
        - $ref: "#/components/parameters/CardNo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "200":
          description: 充值成功，返回收据信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositResponse"
        "400":
          description: 充值金额无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 卡号不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: 卡不是 active 状态（已挂失或已注销）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: CARD_NOT_ACTIVE
                message: 该卡已挂失，无法充值

  # ==================== 就餐消费 ====================
...
  /api/statistics/meal-revenue:
    get:
      tags: [statistics]
      summary: 本餐售饭总收入
      description: 按时间范围统计消费总收入。
      operationId: getMealRevenue
      parameters:
        - $ref: "#/components/parameters/StartTime"
        - $ref: "#/components/parameters/EndTime"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                required: [totalRevenue]
                properties:
                  totalRevenue:
                    type: integer
                    format: int64
                    description: 总收入（分）
  /api/statistics/window-revenue:
    get:
      tags: [statistics]
      summary: 各窗口收入
      description: 按时间范围统计各窗口的消费收入。
      operationId: getWindowRevenue
      parameters:
        - $ref: "#/components/parameters/StartTime"
        - $ref: "#/components/parameters/EndTime"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                required: [windows]
                properties:
                  windows:
                    type: array
                    items:
                      type: object
                      required: [windowId, windowName, revenue]
                      properties:
                        windowId:
                          type: integer
                          format: int64
                        windowName:
                          type: string
                        revenue:
                          type: integer
                          format: int64
                          description: 该窗口收入（分）
  /api/statistics/deposit-details:
    get:
      tags: [statistics]
      summary: 各持卡人存款明细
      description: 按持卡人分组列出每笔存款记录。可选时间范围筛选。
      operationId: getDepositDetails
      parameters:
        - name: startTime
          in: query
          schema:
            type: string
            format: date-time
          description: 起始时间（可选）
        - name: endTime
          in: query
          schema:
            type: string
            format: date-time
          description: 结束时间（可选）
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                required: [holders]
                properties:
                  holders:
                    type: array
                    items:
                      type: object
                      required: [holderId, holderName, idNumber, deposits, totalAmount]
                      properties:
                        holderId:
                          type: integer
                          format: int64
                        holderName:
                          type: string
                        idNumber:
                          type: string
                        deposits:
                          type: array
                          items:
                            type: object
                            required: [id, cardNo, amount, createdAt]
                            properties:
                              id:
                                type: integer
                                format: int64
                              cardNo:
                                type: string
                                description: 16 位卡号
                              amount:
                                type: integer
                                format: int64
                                description: 存款金额（分）
                              createdAt:
                                type: string
                                format: date-time
                        totalAmount:
                          type: integer
                          format: int64
                          description: 该持卡人存款总额（分）
...
  /api/statistics/daily-report:
    get:
      tags: [statistics]
      summary: 日餐报表
      description: 按指定日期汇总当日消费数据。
      operationId: getDailyReport
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: 日期，格式 YYYY-MM-DD
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                required: [date, totalRevenue, transactionCount, windows]
                properties:
                  date:
                    type: string
                    format: date
                  totalRevenue:
                    type: integer
                    format: int64
                    description: 当日消费总额（分）
                  transactionCount:
                    type: integer
                    description: 当日消费笔数
                  windows:
                    type: array
                    items:
                      type: object
                      required: [windowId, windowName, revenue, transactionCount]
                      properties:
                        windowId:
                          type: integer
                          format: int64
                        windowName:
                          type: string
                        revenue:
                          type: integer
                          format: int64
                        transactionCount:
                          type: integer
  /api/statistics/yearly-report:
    get:
      tags: [statistics]
      summary: 年餐报表
      description: 按指定年份汇总每月消费数据。
      operationId: getYearlyReport
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
          description: 年份，如 2026
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                required: [year, totalRevenue, transactionCount, months]
                properties:
                  year:
                    type: integer
                  totalRevenue:
                    type: integer
                    format: int64
                  transactionCount:
                    type: integer
                  months:
                    type: array
                    items:
                      type: object
                      required: [month, revenue, transactionCount]
                      properties:
                        month:
                          type: integer
                          description: 月份（1-12）
                        revenue:
                          type: integer
                          format: int64
                        transactionCount:
                          type: integer

  # ==================== 窗口管理 ====================
...
components:
  parameters:
    CardNo:
      name: cardNo
      in: path
      required: true
      schema:
        type: string
        pattern: '^\d{16}$'
      description: 16 位数字卡号（card_no，非数据库自增 ID）
    StartTime:
      name: startTime
      in: query
      required: true
      schema:
        type: string
        format: date-time
      description: 起始时间（ISO 8601）
    EndTime:
      name: endTime
      in: query
      required: true
      schema:
        type: string
        format: date-time
      description: 结束时间（ISO 8601）
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            错误码枚举：
            - CARD_NOT_FOUND: 卡号不存在
            - CARD_ALREADY_ACTIVE: 该证件号已有 active 卡
            - CARD_NOT_ACTIVE: 卡不是 active 状态
            - CARD_NOT_LOST: 卡不是 lost 状态
            - CARD_CANCELLED: 卡已注销
            - CARD_LOST: 卡已挂失
            - CARD_ALREADY_CANCELLED: 卡已注销，不能重复注销
            - INSUFFICIENT_BALANCE: 余额不足
            - INVALID_AMOUNT: 金额无效
            - WINDOW_NOT_FOUND: 窗口不存在
            - VALIDATION_ERROR: 通用校验错误
            - STUDENT_NOT_FOUND: 证件号不存在（非本校人员）
            - STUDENT_SERVICE_ERROR: 学籍验证服务不可用
            - INTERNAL_ERROR: 服务端内部错误
        message:
          type: string

    # ---------- 学籍验证 ----------
...
    DepositRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          format: int64
          minimum: 1
          description: 充值金额（分），必须大于 0
    DepositResponse:
      description: 充值成功的收据信息
      type: object
      required: [id, cardNo, holderName, amount, newBalance, createdAt]
      properties:
        id:
          type: integer
          format: int64
          description: 存款记录 ID
        cardNo:
          type: string
          description: 16 位卡号
        holderName:
          type: string
        amount:
          type: integer
          format: int64
          description: 充值金额（分）
        newBalance:
          type: integer
          format: int64
          description: 充值后余额（分）
        createdAt:
          type: string
          format: date-time

    # ---------- 就餐消费 ----------
...